Ft frontend integration fixes (#28)
* FIxed login

* FIxed cart

+++ src/modules/cart/repository/cartRepository.js
@@ -22,8 +22,13 @@ const getCartByAttributes = async (key1, value1, key2, value2) => {
   return await Cart.findOne(query);
 };
 
-const deleteProductFromCart = async (productId) => {
-  return await Cart.findOneAndDelete({ productId })
+const findCartByProductAndUserId = async (productId, userId) => {
+  return await Cart.findOne({ productId, userId });
+
+}
+
+const deleteProductFromCart = async (productId, userId) => {
+  return await Cart.findOneAndDelete({ productId, userId });
 }
 
 const findCartItemsByUserId = async (userId) => {
@@ -37,5 +42,6 @@ export default {
   getCartByAttribute,
   getCartByAttributes,
   deleteProductFromCart,
-  findCartItemsByUserId
+  findCartItemsByUserId,
+  findCartByProductAndUserId
 };
